Updated SOFTWARE/Control
The first version didn't work - it blinked randomly while a control signal was applied, and with a constant frequency when the control signal was removed.

The program was restructured to run entirely in the ISR, and the startup check and timer1 overflow error was removed. More documentation was also added.